pyproject.toml
[project]
name = "oddo-zipcodebreakers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.6",
    "bcrypt==4.0.1",
    "databases>=0.9.0",
    "fastapi>=0.127.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.40.0",
]